+++ NEWS.md
@@ -1,3 +1,4 @@
 # nlmixr2targets 0.0.0.9000
 
 * Added "cens" and "limit" columns as nlmixr2 columns
+* Keep columns from the `table` argument's keep element